Moved the stylesheets to its own folder.
